Update TestInteractions.java
Forgot comment on previous commit:

Edited second archer interaction test to be based on a 99% confidence interval. See below for the stat-sy details:

(Note: not at all your fault Ethan as I literally just fixed some things with archer that would have affected this anyway)

The probability of getting hit assuming the player is at location 1 and the archer is at location 15 (they are 14 away from each other) is (21-14)/42 = 7/42. Based on some stats stuff (bootstrap distribution), for a sample size of 500 the number of hits should fall between 57 and 109 99% of the time